Update scrape.py
Added a function that checks if the following link has multiple topics or goes straight to a forum.

I am so sick of this project but I must get it done! The dopamine from overcoming the obstacles is getting to me.

It feels weird not to use functional programming, but I think that this is much easier.

Also, what a NOT pythonic way to write this right? But if it works it works. It's doing what the recursive version did except 3x faster and with half as much code.