Update scrape.py
Okay, this should be the final version. I will run it overnight and see if it fails.

This time I added implementation for the other menu variations. I wrapped the forum into a function to prevent repeating code. However, this meant that I needed to implement a path parameter, which was no problem.

The next step I would like to take if this works, is creating a folder for each forum. Inside this folder, will have downloaded and labeled images, as well as a copy of the raw HTML code. However, this will likely take up 2-3x more storage and is unnecessary.