Update scrape.py
As expected, I ran into some issues. For some reason, random posts appear every 10,000 or so that do not have the proper classes in them. I wrapped my function in a try, except because I can not locate these posts. In addition, I also added a delay. I found that the server was automatically blocking me because I was sending out too many requests, and through trial and error found out that 400 milliseconds between each request + the processing time is enough to keep the server from kicking me off. 

The trial and error process is taking FOREVER. I run into an error every 4-5 hours or so, fix it, and move on. I am thinking of implementing a way to check if the script is still running, and if it is not, to send me an email and play a really loud noise. Just thought, though.

While I am waiting, I am going to begin to implement a way to capture images. Many of the posts are referring to images and it makes it takes away from the reading.