Update scrape.py
I wrapped everything in try-excepts because I didn't want to waste any time. I initially tinkered with creating an error file to save, but didn't want to implement saving and writing to another file. This program has slowly become spaghetti code! My next steps are to clean it up a bit, if I can, and then to implement further features, then I will be done. All I want to do is add photos. I also had to add a feature to preserve the link on one of the three branch possibilities because it would cause nasty errors.